pyproject.toml
[project]
name = "voice-gen"
version = "0.1.0"
description = "Text to audio generator"
authors = [{ name = "Alex Nal" }]
readme = "README.md"
requires-python = ">=3.9,<3.12"
dependencies = ["coqui-tts"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
